+++ scripts/build.py
@@ -85,7 +85,7 @@ def configure_cmake(args : argparse.Namespace) -> None:
         elif args.arch == "arm":
             cmake_args.append("arm")
         # No support for ARM32 right now
-    
+
     if args.host_arch != args.arch:
         cmake_args.append("-DSYMCRYPT_TARGET_ARCH=" + args.arch.upper())
 
@@ -109,6 +109,17 @@ def configure_cmake(args : argparse.Namespace) -> None:
 
     if args.verbose:
         cmake_args.append("-DCMAKE_VERBOSE_MAKEFILE=ON")
+
+    if args.openssl_build_from_source:
+        cmake_args.append("-DOPENSSL_BUILD_FROM_SOURCE=ON")
+        args.openssl = True
+
+    if args.openssl:
+        cmake_args.append("-DSYMCRYPT_TEST_WITH_OPENSSL=ON")
+
+    if args.openssl_branch:
+        cmake_args.append("-DOPENSSL_BUILD_BRANCH=" + args.openssl_branch)
+
 
     if args.clean and args.build_dir.exists():
         shutil.rmtree(args.build_dir)
@@ -203,6 +214,9 @@ def main() -> None:
     parser_cmake.add_argument("--configure-only", action = "store_true", help = "Run CMake configuration, but do not build.")
     parser_cmake.add_argument("--no-parallel-build", action = "store_false", dest = "parallel_build", help = "Disable parallel CMake build.", default = True)
     parser_cmake.add_argument("--verbose", action = "store_true", help = "Enable CMake verbose mode.", default = False)
+    parser_cmake.add_argument("--openssl", action = "store_true", help = "Enable OpenSSL performance comparison.", default = False)
+    parser_cmake.add_argument("--openssl-branch", type = str, help = "Checkout and build specified branch of OpenSSL.", default = None)
+    parser_cmake.add_argument("--openssl-build-from-source", action = "store_true", help = "Build OpenSSL from source.", default = False)
 
     # MSBuild build options
     parser_msbuild = subparsers.add_parser("msbuild", help = "Build using MSBuild.")