+++ unittest/inc/perf.h
@@ -1,7 +1,7 @@
 //
 // Performance measuring infrastructure header
 //
-// Copyright (c) Microsoft Corporation. Licensed under the MIT license. 
+// Copyright (c) Microsoft Corporation. Licensed under the MIT license.
 //
 
 
@@ -89,10 +89,10 @@
 #endif
 
 //
-// The perf measuring is complicated. The P4 has all kind of alignment issues; 
+// The perf measuring is complicated. The P4 has all kind of alignment issues;
 // The L1 cache can't hold two cache lines that are aliased mod 64 kB. This leads
 // to huge perf differences that are all bogus if you try to compare implementations.
-// 
+//
 // We solve this by running measurements using many different addresses.
 // Each perf function can use up to four memory addresses, each of which is at least
 // one MB large. We chose random pointer addresses and run the tests many times.
@@ -100,7 +100,7 @@
 //
 
 //
-// For algorithms that specify a keying function the keying function will be called before the data 
+// For algorithms that specify a keying function the keying function will be called before the data
 // function. The keying function should set up any necessary expanded key.
 //