+++ unittest/lib/main_exe_common.cpp
@@ -1,18 +1,18 @@
 //
-// Copyright (c) Microsoft Corporation. Licensed under the MIT license. 
+// Copyright (c) Microsoft Corporation. Licensed under the MIT license.
 //
 
-const char * g_implementationNames[] = 
+const char * g_implementationNames[] =
 {
     ImpSc::name,
-#if INCLUDE_IMPL_RSA32    
+#if INCLUDE_IMPL_RSA32
     ImpRsa32::name,
     ImpRsa32b::name,
 #endif
-#if INCLUDE_IMPL_CAPI    
+#if INCLUDE_IMPL_CAPI
     ImpCapi::name,
 #endif
-#if INCLUDE_IMPL_CNG    
+#if INCLUDE_IMPL_CNG
     ImpCng::name,
 #endif
 #if INCLUDE_IMPL_REF
@@ -20,11 +20,14 @@ const char * g_implementationNames[] =
 #endif
 #if INCLUDE_IMPL_MSBIGNUM
     ImpMsBignum::name,
-#endif    
+#endif
+#if INCLUDE_IMPL_OPENSSL
+    ImpOpenssl::name,
+#endif
     NULL,
 };
 
-VOID 
+VOID
 addAllAlgs()
 {
     addSymCryptAlgs();
@@ -32,14 +35,14 @@ addAllAlgs()
     if( !g_sgx )
     {
         // SGX mode cares about testing BCrypt functions in enclaves, so ignores CAPI and RSA32 tests
-        // which are identical to normal mode. SymCrypt provides implementations for all algs, 
+        // which are identical to normal mode. SymCrypt provides implementations for all algs,
         // so they must run because the test fails if there are algorithms where no implementations were tested.
 #if INCLUDE_IMPL_RSA32
         addRsa32Algs();
 #endif
 #if INCLUDE_IMPL_CAPI
         addCapiAlgs();
-#endif            
+#endif
     }
 
 #if INCLUDE_IMPL_CNG
@@ -50,7 +53,10 @@ addAllAlgs()
 #endif
 #if INCLUDE_IMPL_MSBIGNUM
     addMsBignumAlgs();
-#endif        
+#endif
+#if INCLUDE_IMPL_OPENSSL
+    addOpensslAlgs();
+#endif
 }
 
 int SYMCRYPT_CDECL